requirements.txt
litellm>=1.35.9
chromadb>=0.5.0
pydantic>=2.7
pyyaml>=6.0.1
typer>=0.12.3
rich>=13.7
sqlite-utils>=3.36
pytest>=8.2
